Refactor asciidoc documentation
Create the `asciidoc` package and separate the command line reference
documentation generation into the `asciidoc/cli` package.

Reference: EC-513
