+++ internal/documentation/asciidoc/cli/cli.template
